KS-1-4 task4 work with types_refactored

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -4,9 +4,9 @@ fun main() {
     val length: Long = 40_868_600_000
     val age: Byte = 27
     val partOfDay: Float = 0.075f
-    val seconds: UInt = 6480u
+    val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
-    val apogee: UInt = 327000u
+    val apogee: Int = 327000
 
     println("\n$length\n$age\n$partOfDay\n$seconds\n$partOfYear\n$apogee")
 }
